remove duplicate xlib-gl bridge and move glew init to bridge (#1317)
Move glew initialization to bridges completely and cleanup and deduplicate Win32 and Xlib bridges for OpenGL graphics systems.
