+++ ENIGMAsystem/SHELL/Bridges/Win32-OpenGL1/Makefile
@@ -1,1 +1,1 @@
-SOURCES += $(wildcard Bridges/Win32-OpenGL1/*.cpp)
+SOURCES += $(wildcard Bridges/Win32-OpenGL1/*.cpp) $(wildcard Bridges/Win32-OpenGL/*.cpp)
