+++ ENIGMAsystem/SHELL/Bridges/Win32-OpenGL1/graphics_bridge.cpp
@@ -25,163 +25,60 @@
 **                                                                              **
 \********************************************************************************/
 
-#include <string>
-#include <windows.h>
-#include <GL/glew.h>
-#include <GL/wglew.h>
-
-using namespace std;
-
-#include "libEGMstd.h"
 #include "Widget_Systems/widgets_mandatory.h"
 #include "Platforms/Win32/WINDOWSmain.h"
 #include "Platforms/General/PFwindow.h"
 #include "Graphics_Systems/General/GScolors.h"
 
+#include <windows.h>
+#include <GL/glew.h>
+#include <GL/wglew.h>
+
 namespace enigma
 {
-    GLuint msaa_fbo = 0;
 
-    extern void (*WindowResizedCallback)();
-    void WindowResized() {
-      // clear the window color, viewport does not need set because backbuffer was just recreated
-      enigma_user::draw_clear(enigma_user::window_get_color());
-    }
-
-    extern HGLRC hRC;
-
-    void EnableDrawing(void*)
-    {
-      WindowResizedCallback = &WindowResized;
-
-      PIXELFORMATDESCRIPTOR pfd;
-      int iFormat;
-
-      enigma::window_hDC = GetDC (hWnd);
-      ZeroMemory (&pfd, sizeof (pfd));
-      pfd.nSize = sizeof (pfd);
-      pfd.nVersion = 1;
-      pfd.dwFlags = PFD_DRAW_TO_WINDOW | PFD_SUPPORT_OPENGL | PFD_DOUBLEBUFFER;
-      pfd.iPixelType = PFD_TYPE_RGBA;
-      pfd.cColorBits = 24;
-      pfd.cDepthBits = 24;
-      pfd.cStencilBits = 8;
-      pfd.iLayerType = PFD_MAIN_PLANE;
-      iFormat = ChoosePixelFormat (enigma::window_hDC, &pfd);
-
-      if (iFormat==0) { show_error("Failed to set the format of the OpenGL graphics device.",1); }
-
-      SetPixelFormat (enigma::window_hDC, iFormat, &pfd);
-      hRC = wglCreateContext( enigma::window_hDC );
-      wglMakeCurrent( enigma::window_hDC, hRC );
-
-      //TODO: This never reports higher than 8, but display_aa should be 14 if 2,4,and 8 are supported and 8 only when only 8 is supported
-      glGetIntegerv(GL_MAX_SAMPLES_EXT, &enigma_user::display_aa);
-    }
-
-    void DisableDrawing(void*)
-    {
-      wglMakeCurrent (NULL, NULL);
-      wglDeleteContext (hRC);
-      ReleaseDC (enigma::hWnd, enigma::window_hDC);
-    }
+extern void (*WindowResizedCallback)();
+void WindowResized() {
+  // clear the window color, viewport does not need set because backbuffer was just recreated
+  enigma_user::draw_clear(enigma_user::window_get_color());
 }
 
-// NOTE: Changes/fixes that applies to this likely also applies to the OpenGL1 version.
+extern HGLRC hRC;
 
-namespace enigma {
-  namespace swaphandling {
-    bool has_checked_extensions = false;
-    bool ext_swapcontrol_supported;
+void EnableDrawing(void*)
+{
+  WindowResizedCallback = &WindowResized;
 
-    void investigate_swapcontrol_support() {
+  PIXELFORMATDESCRIPTOR pfd;
+  int iFormat;
 
-      if (has_checked_extensions) return; // Already calculated, no need to calculate it more.
+  enigma::window_hDC = GetDC (hWnd);
+  ZeroMemory (&pfd, sizeof (pfd));
+  pfd.nSize = sizeof (pfd);
+  pfd.nVersion = 1;
+  pfd.dwFlags = PFD_DRAW_TO_WINDOW | PFD_SUPPORT_OPENGL | PFD_DOUBLEBUFFER;
+  pfd.iPixelType = PFD_TYPE_RGBA;
+  pfd.cColorBits = 24;
+  pfd.cDepthBits = 24;
+  pfd.cStencilBits = 8;
+  pfd.iLayerType = PFD_MAIN_PLANE;
+  iFormat = ChoosePixelFormat (enigma::window_hDC, &pfd);
 
-      const char *wgl_extensions = wglGetExtensionsStringARB(window_hDC);
+  if (iFormat==0) { show_error("Failed to set the format of the OpenGL graphics device.",1); }
 
-      ext_swapcontrol_supported = strstr(wgl_extensions, "WGL_EXT_swap_control");
+  SetPixelFormat (enigma::window_hDC, iFormat, &pfd);
+  hRC = wglCreateContext( enigma::window_hDC );
+  wglMakeCurrent( enigma::window_hDC, hRC );
 
-      has_checked_extensions = true;
-    }
+  // -- Initialise GLEW
+  GLenum err = glewInit();
+  if (GLEW_OK != err)
+  {
+    return;
   }
 
-  bool is_ext_swapcontrol_supported() {
-    swaphandling::investigate_swapcontrol_support();
-    return swaphandling::ext_swapcontrol_supported;
-  }
+  //TODO: This never reports higher than 8, but display_aa should be 14 if 2,4,and 8 are supported and 8 only when only 8 is supported
+  glGetIntegerv(GL_MAX_SAMPLES_EXT, &enigma_user::display_aa);
 }
 
-#include "Universal_System/roomsystem.h"
-
-namespace enigma_user {
-
-	int display_aa = 0;
-
-	void display_reset(int samples, bool vsync) {
-		int interval = vsync ? 1 : 0;
-
-		if (enigma::is_ext_swapcontrol_supported()) {
-		  wglSwapIntervalEXT(interval);
-		}
-
-		GLint fbo;
-		glGetIntegerv(GL_FRAMEBUFFER_BINDING_EXT, &fbo);
-
-		GLuint ColorBufferID, DepthBufferID;
-
-		// Cleanup the multi-sampler fbo if turning off multi-sampling
-		if (samples == 0) {
-			if (enigma::msaa_fbo != 0) {
-				glDeleteFramebuffers(1, &enigma::msaa_fbo);
-				enigma::msaa_fbo = 0;
-			}
-			return;
-		}
-
-		//TODO: Change the code below to fix this to size properly to views
-		// If we don't already have a multi-sample fbo then create one
-		if (enigma::msaa_fbo == 0) {
-			glGenFramebuffersEXT(1, &enigma::msaa_fbo);
-		}
-		glBindFramebufferEXT(GL_FRAMEBUFFER_EXT, enigma::msaa_fbo);
-		// Now make a multi-sample color buffer
-		glGenRenderbuffersEXT(1, &ColorBufferID);
-		glBindRenderbufferEXT(GL_RENDERBUFFER_EXT, ColorBufferID);
-		glRenderbufferStorageMultisampleEXT(GL_RENDERBUFFER_EXT, samples, GL_RGBA8, window_get_region_width(), window_get_region_height());
-		// We also need a depth buffer
-		glGenRenderbuffersEXT(1, &DepthBufferID);
-		glBindRenderbufferEXT(GL_RENDERBUFFER_EXT, DepthBufferID);
-		glRenderbufferStorageMultisampleEXT(GL_RENDERBUFFER_EXT, samples, GL_DEPTH_COMPONENT24, window_get_region_width(), window_get_region_height());
-		// Attach the render buffers to the multi-sampler fbo
-		glFramebufferRenderbufferEXT(GL_FRAMEBUFFER_EXT, GL_COLOR_ATTACHMENT0_EXT, GL_RENDERBUFFER_EXT, ColorBufferID);
-		glFramebufferRenderbufferEXT(GL_FRAMEBUFFER_EXT, GL_DEPTH_ATTACHMENT_EXT, GL_RENDERBUFFER_EXT, DepthBufferID);
-
-	}
-
-  void screen_refresh() {
-    window_set_caption(room_caption);
-    enigma::update_mouse_variables();
-    SwapBuffers(enigma::window_hDC);
-  }
-
-  void set_synchronization(bool enable) {
-
-    // General notes:
-    // Setting swapping on and off is platform-dependent and requires platform-specific extensions.
-    // Platform-specific extensions are even more bothersome than regular extensions.
-    // What functions and features to use depends on which version of OpenGL is used.
-    // For more information, see the following pages:
-    // http://www.opengl.org/wiki/Load_OpenGL_Functions
-    // http://www.opengl.org/wiki/OpenGL_Loading_Library
-    // http://www.opengl.org/wiki/Swap_Interval
-    // http://en.wikipedia.org/wiki/WGL_%28software%29
-    // Also note that OpenGL version >= 3.0 does not use glGetString for getting extensions.
-
-      int interval = enable ? 1 : 0;
-
-    if (enigma::is_ext_swapcontrol_supported()) {
-      wglSwapIntervalEXT(interval);
-    }
-  }
 }
