+++ ENIGMAsystem/SHELL/Bridges/Win32-OpenGL3/Makefile
@@ -1,1 +1,1 @@
-SOURCES += $(wildcard Bridges/Win32-OpenGL3/*.cpp)
+SOURCES += $(wildcard Bridges/Win32-OpenGL3/*.cpp) $(wildcard Bridges/Win32-OpenGL/*.cpp)
