+++ ENIGMAsystem/SHELL/Graphics_Systems/OpenGL3/OPENGL3Std.cpp
@@ -53,15 +53,6 @@ namespace enigma
   void graphicssystem_initialize()
   {
     oglmgr = new ContextManager();
-    #ifdef DEBUG_MODE
-    GLenum err = glewInit();
-    if (GLEW_OK != err)
-    {
-      std::cout<<"GLEW ERROR!"<<std::endl;
-    }
-    #else
-    glewInit();
-    #endif
 
     glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
